requirements.txt
torch
transformers
Pillow
requests
huggingface_hub
python-dotenv
accelerate